Рабочий билд 0.9 / Скажем нет не авторизованным пользователям! / Ограничение доступа к graphql запросам по авторизации и роли / Замена штатного AppoloProvider

+++ backend/app/GraphQL/Mutations/DeleteContact.php
@@ -2,12 +2,22 @@
 
 namespace App\GraphQL\Mutations;
 
+use App\GraphQL\Service\AuthorizationService;
 use App\Models\Contact;
+use Nuwave\Lighthouse\Exceptions\AuthenticationException;
+use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 
 final readonly class DeleteContact
 {
-    public function __invoke($_, array $args)
+    public function __invoke($_, array $args, GraphQLContext $context)
     {
-        $contact = Contact::destroy($args['id']);
+        $allowedRoles = ['admin']; // Роли, которые разрешены
+        $accessToken = $context->request()->header('Authorization');
+        if (AuthorizationService::checkAuthorization($accessToken, $allowedRoles)) {
+            $contact = Contact::destroy($args['id']);
+        } else {
+            throw new AuthenticationException('Отказано в доступе');
+        }
+
     }
 }
