+++ backend/app/GraphQL/Mutations/DeleteOrganization.php
@@ -2,14 +2,24 @@
 
 namespace App\GraphQL\Mutations;
 
+use App\GraphQL\Service\AuthorizationService;
 use App\Models\Organization;
+use Nuwave\Lighthouse\Exceptions\AuthenticationException;
+use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 
 final readonly class DeleteOrganization
 {
     /** @param  array{}  $args */
-    public function __invoke(null $_, array $args)
+    public function __invoke(null $_, array $args, GraphQLContext $context)
     {
-        $organization = Organization::destroy($args['id']);
+        $allowedRoles = ['admin']; // Роли, которые разрешены
+        $accessToken = $context->request()->header('Authorization');
+        if (AuthorizationService::checkAuthorization($accessToken, $allowedRoles)) {
+            $organization = Organization::destroy($args['id']);
+        } else {
+            throw new AuthenticationException('Отказано в доступе');
+        }
+
 
     }
 }
